consumer/Weather/requirements.txt
kafka-python
psycopg2-binary
python-dotenv